Add a function to change the minter identity
It could be useful to transfer the privilege to somebody else, for example when the minter leaves a job with the organization that owns the ERC-721 contract. This could also be accomplished by giving the private key for the address to the next person, but if a person created multiple ERC-721 contracts and didn't think about it beforehand they might have to give all of them or none. With this function they can decide to transfer only some contracts.