This consolidates and clarifies top level module names.
functions/ only dealt with builtin functions so it's been renamed.
interfaces/ only dealt with builtin interfaces so it's been renamed.

codegen/ was going to be a module to replace parser, but we are going to
replace parser/ entirely with another pass (name tbd), so throw it all
in together.

signatures/ had too much ast code so it went in ast/.

settings.py seemed like it belonged in compiler/
opcodes seemed like it belonged in a new folder for evm specifics.
compile_lll, optimizer and s_expressions seemed to belong in a new lll/
folder

context/ was moved into ast/ since it has a bunch of type and
environment info which belongs in the AST pass. I think the intent was
to have ast/ and context be separate passes (so that parsing and
validation occurred in separate passes) but because they all deal with
the same representation I decided to merge them into the same module and
mark validation as a separate pass (submodule). The entire context/
folder was flattened so that:

context/types/ moved up a relative level into ast/types/
context/validation/ moved up a relative level into ast/validation/
context/namespace.py and context/environment.py moved up to the ast/
module.

move types/ into parser/ since it is the old types
move parser/ into old_codegen/ to mark it as an old module which are
getting rid of.