fix the test for storage slot
The last storage variable was not being tested which is why this slipped
by.

The reason there was a collision in the last storage variable is because
there are holes in the storage allocator left by HashMaps (which is why
the slots in global_context and set_data_positions were not matching
up).