Fix return optimization for nested tuple
The optimization for tuples of base types expected anything inside a
tuple to be an int, but that's not always the case.

Luckily logic introduced in 0605abbe9 can actually be generalized to any
tuple type which has no dynamic data in the ABI encoding, so we just
remove the constraints on the child types.