Implement user-facing abi encode function
This exposes _abi_encode to the user. The returned bytestring has a
length of the size bound on the ABI encoded data (size_bound was
refactored in this commit to a convenience function) and the provided
buffer must be large enough to accept the data.

re. naming:
We haven't figured out the namespace, so use an underscore to suggest
the name is provisional.