Implement abi_encode builtin (#2401)
* Implement user-facing abi encode function

This exposes _abi_encode to the user. The returned bytestring has a
length of the size bound on the ABI encoded data (size_bound was
refactored in this commit to a convenience function) and the provided
buffer must be large enough to accept the data.

re. naming:
We haven't figured out the namespace, so use an underscore to suggest
the name is provisional.

* Add ensure_tuple kwarg to _abi_encode

`ensure_tuple` toggles the behavior of whether to force single values into
a tuple or not. The default is True (same as solidity)

Also add tests, fix some bugs

Co-authored-by: El De-dog-lo <3859395+fubuloubu@users.noreply.github.com>

Co-authored-by: El De-dog-lo <3859395+fubuloubu@users.noreply.github.com>