fix: allow start index to be determined at runtime slicing msg.data
Since the length of msg.data is dynamic, we can't say for sure what it's
maxlength is. Even inside of a function with 3 word length arguments,
data can still be appended to a function's calldata. So we make length
a compile time requirement when using msg.data and slice, and we allow
start to be determined at runtime since we don't need it to determine
the size of the byte array.