Prune inefficient jumps
The new function codegen generates a lot of code like
```
goto function_common
label function_common
```

This prunes that code.

It also prunes labels that are never used, e.g. there is a join point at
the end of all functions which is like
```
label function__cleanup
```

If there is no return statement in the function the label is
unnecessary.