fix: account for immutables at end of runtime code
Immutable values are appended to the runtime code, and so using
`codesize` opcode will include them in the returned size. However, the
immutables section of the runtime code is not at the end, so to access
it we need to account for the fact that `codesize` gives us an inflated
value. We do this by taking the sum of all the allocated space of all
immutables, and subtrace codesize by this and then add the offset of the
data.