fix version string generated in docker builds
Dockerhub changes the executable bit on `hooks/build`, which leads to
the repository being dirty during the build which messes with the
version string created by setuptools-scm. This commit changes the
executable bit so that the docker build does not make the repo dirty.

It also adds a `git reset --hard` to ensure the repo is clean even if
the build process dirties the repo in a different way in the future.