feat: implement dynamic arrays
This commit implements Solidity-ABI-compatible dynamic arrays in vyper.

Dynamic arrays have the type `DynArray[<type>, <maxlen>]`. They are
implemented similarly to bytestrings, with a single length word followed
by the data, <len> <data...>.

It also adds some code to clarify between the sizes of objects
when they are in memory vs in storage.

This commit also refactors a lot of code dealing with encoding and byte
copying. Byte copying is cheaper now since previously the logic was
```
for i in range(maxlen // 32):
    if i > len:
        break
    sstore(dst + i, mload(src + 32*i))
```

This was necessary because the `repeat` opcode in LLL requires a max
bound. Now, we have repeat(len, maxlen), which still enforces the max
bound but breaks once it hits the minimum of `len` and `maxlen`. Now the
byte copier loop has the logic of:

```
for i in range( min(len, maxlen) ):
    sstore(dst + i, mload(src + 32*i))
```

It also adds a method to the LLLnode API, `cache_when_complex`. This
abstracts the logic for checking when an LLL expression needs to be
cached using a `with` expression. This was used to simplify the ABI
encoder and several routines related to copying (including
`make_setter`).