re-enable msg.* env variables in internal functions (#2632)
There used to be confusion as to the value of `msg.*` when we           
implemented internal functions as call-to-self, so we disabled `msg.*`  
in internal functions in #1199. However, since then internal functions  
are implemented just using JUMPs, so the EVM execution context doesn't  
change. In other words, msg.sender is the same no matter if it's used in
an internal or external function, so we can lift this restriction.

fix #2631