chore: refactor dynamic arrays code (#2633)
This commit fixes some performance bugs in dynamic arrays.         
                                
It uses a loop to copy dynamic arrays when the child type is also  
dynamic (can take less space than is allocated for it).            
                                
It also moves the loop variable in the `repeat` macro to the stack,
reducing memory traffic in loops.
                                
It also fixes a bug in the `in` operator where side effects in the LHS
can get repeatedly evaluated.