feat: add parametrized labels to LLL (#2598)
Change labels in LLL to be parameterized. Previously, stack items were
implicitly consumed from jumpdests using `pass`, which made liveness
analysis for stack items impossible in general. Now, any parameters
passed on the stack are expicitly named. Since labels now take
parameters, they are very similar to subroutines so we could use that
terminology interchangeably. Note that this changes the structure of
labels - instead of allowing labels to occur arbitrarily in code, they
turn into blocks, enforcing scope of the named stack items.

This change also enables translation to other architectures such as LLVM
backends since implicit stack items cannot be handled in translation to
a register machine.

This commit also renames `get_comb_opcodes` to `get_lll_opcodes`.
It originally stood for "get combined [pseudo- and evm-] opcodes". What
it really meant is "get all LLL opcodes".

Lastly, this commit also adds an `exit_to` instruction for LLL. This
compiles the same as `goto` -- to a JUMP instruction -- but has
different scoping semantics. `exit_to` terminates a subroutine while
`goto` expects control flow to be returned. This is an important hint
for liveness analysis. If you hit an `exit_to`, all stack items in scope
must be consumed, whereas if you hit a `goto` the same requirement does
not hold.

In order to accomodate the semantic change, this commit changes the
codegen everywhere that stack items are consumed implicitly. In
practice, this happens in private functions, and in function common
cleanup blocks. Also, all blocks must now be terminated (either with an
EVM stopping instruction like STOP, or with the new `exit_to` macro).

Note: the assembly output for all contracts I looked at were
byte-for-byte the same with the exception of those which returned
straight out of a loop (using the LLL `cleanup_repeat` instruction), so
this commit adds a couple more tests to test returns out of loops (in
the void return and external cases)