feat: add `select` opcode to LLL
this is an optimizing opcode for ternary operations. this is useful
because even in the presence of perfect liveness analysis, branches
still need to issue more POPs than are required with a specialized
`select` opcode.

this commit also modifies the `min/max` builtins to use the new opcode,
as an example.