feat: add all integer and bytes types
This commit adds all intN/uintN/bytesM types, bringing Vyper to full
implementation of ABIv2 (technically in line with solidity as they do
not have fixedMxN/ufixedMxN types). Practically speaking, these are the
types that people will expect to be able to use.

The implementation is a first pass. It generates the new types with
Python dynamic class generation since the vyper/semantics type system
has a 1:1 correspondence between Vyper and Python types. Future work
should include a generous refactoring of the approach. Before release,
we need to add more tests of arithmetic, and any builtin functions which
accept integers and bytesM types.