fix: erc721 erc165 support (#2622)
Currently, the example in the Vyper org for ERC721 is not 100% compliant
because it doesn't support the ERC165 method_id  
supportsInterface(bytes4). This commit works around by using the
signature pizza_mandate_apology(uint256) which has the same method ID as
supportsInterface(bytes4). Future work is to update to use the new
bytes4 type.

Co-authored-by: Charles Cooper <cooper.charles.m@gmail.com>