fix: allow struct constants (#2617)
This commit enables constant structs, and constant struct members. 
It also introduces a new helper function named "check_kwargable",
that checks if a given node can be a kwarg to a function. This is 
meant to clarify the existing "check_constant" helper function, 
which actually checks whether a given node is not assignable.

The fix for constant structs is handled in the AST folding by
replacing references to a constant struct with the corresponding
Call AST node. The constant struct is then recursively checked
by the "check_constant" helper function to ensure all nested
structs and struct members are not assignable or a literal value.

This commit also involves a related change to the semantics 
validation of a function. Previously, both the validation of a 
function's kwargs and a module-level constant uses the same
"check_constant" function. While it previously conveniently 
covered both use cases of "not_assignable" and "is_kwargable",
this approach breaks for constant structs. Therefore, a new 
helper function "check_kwargable" has been added to separate
the use cases.

Resolves: #1430 
See also: #2396