optimize byte array copiers
this commit optimizes the make_byte_array_copier and
make_dynarray_copier routines. specifically, it reduces the number of
words to copy in the main loop by 1. for batch copies
(memory/calldata/code -> memory), this reduces gas by 3, and for
non-batch copies (e.g. storage -> memory), this reduces gas by one loop
iteration and one read. importantly, when the number of bytes is small
(e.g. copy Bytes[32]), this triggers an optimization in copy_bytes so
the whole copy operation is a single load/store.