refactor conversion tests (#2756)
refactor conversion tests to cover more edge cases by rewriting the test   
cases around a core "spec" function, _py_convert, which implements the
conversion rules on python objects. this could potentially be moved to
the main codebase in the future, to help with literal handling and also
to serve as a reference implementation for the conversion rules.     

also fix some bugs in conversion:  
- minor bugs in literal handling   
- clamps for address -> int used the clamp rules for int -> address  
- update the bounds for conversions to decimal    

this commit also factors out the unsigned_to_signed utility function 
from test_unsafe_math.py. I found that an alternative, "forgiving"   
implementation (as below) led to bugs downstream since it allows     
consumers of the function to not sanitize inputs. so I switched to the
unforgiving implementation.

```          
def unsigned_to_signed(int_, bits):               
    int_ %= (2 ** (bits - 1) - 1:  # wrap the input
    if int_ > 0:         
        int_ -= 2 ** bits 
    return int
```