fix: package version (#2816)
this is a hybrid approach. previously we used bumpversion to bump the
version and then appended the commithash to the `--version` output. as
of 0.3.1, we used setuptools-scm to append the commit hash. however,
setuptools-scm did not have a way to create multiple versions -- one for
our `--version` output, and one for PKG-INFO.

this commit uses setuptools-scm to automate the global (ex. v0.3.2)
version string calculation, while using an auxiliary file to keep track
of the commit hash for `--version`.

also, add old version files to .gitignore and `make clean`.

even though we don't generate them anymore, they sometimes get generated
when building old releases locally