feat: publish immutables layout (#2768)
add immutables layout to -f layout. this is a breaking change to the
`layout` output, since it adds a top-level key to distinguish between
storage and immutable variables.

this also blocks immutables from colliding with storage variables.
immutables could have the same name as storage variables, depending on
the order they occur in. for instance, this was allowed
```
x: uint256
x: immutable(uint256)
```

but, this was not:
```
x: immutable(uint256)
x: uint256
```

this commit ensures both cases are blocked.