refactor IR optimizer
this is a small rewrite of the IR optimizer. it changes the structure of
the binop optimizations so that it is easier to add more optimizations.
it also refactors the `clamp` optimizations to be in terms of an
`assert` statement, so that the clamp conditions can be optimized using
the binop optimizer code.