loop assembly optimizer
some optimizations in assembly are not straight-line. that is, if step B
comes after step A, A could open up optimizations for and B could open
up optimizations for A. this is especially true for the jumpdest pruner
steps. this commit loops the assembly optimizer steps so that they are
not as ordering dependent.