feat: call internal functions from ctor
this commit allows the user to call internal functions from the
`__init__` function. it does this by tracing all the internal functions
called from every function's execution (during the annotation phase),
and then generating code for the functions called from the init function
for during deploy code generation

this also has a performance benefit (compiler time) because we can get
rid of the two-pass method for tracing frame size.

also, this opens the door to topsort of internal functions, if desired
(so that the user does not have to order them manually).