feat: add `default_return_value` kwarg for calls (#2839)
this commit adds a new kwarg, `default_return_value`, available to all
external calls. its purpose is to make it easier to interface with
contracts which are not compliant with their interfaces (e.g., bad ERC20
implementations) without having to drop down to `raw_call`.

since the logic for handling kwargs in `external_call.py` was getting a
little gnarly, this commit also refactors and that module a bit. for
consistency, it also switches to exclusively using the
typechecker type of the function.