allow msg.data in raw_call without slice
a common use case for raw_call is to forward all calldata to an
implementation or logic contract of some kind. this enables that without
requiring the length of msg.data to be known. it does this by simply
copying calldata to the location pointed to by `msize`.

this commit also fixes an annotation in abi_decode, and slightly cleans
up the logic in the raw_call implementation.