feat: optimize initcode size
if initcode is going to be deployed as part of a factory pattern, we
need to try to optimize the codesize of initcode (which previously was
not a priority, since codesize is only constrained for runtime code).

this commit optimizes PUSH instructions for memory symbols -- memory
locations of immutable variables. it does this by assembling the runtime
code as the first pass in assembly_to_evm, and then using the smallest
possible PUSH instruction which will handle all the memory symbols that
will be generated.