feat: arithmetic for new int types (#2843)
this commit generalizes codegen for arithmetic ops to handle any
integer type. to do so, it refactors codegen for arithmetic ops into its
own module. it also slightly changes the codegen so that it targets
optimizer rules better.

this commit also changes DecimalInfo.divisor to be of `int` type. this
is slightly less convenient, but it's clearer when used that it is not
getting truncated.

also, fix a pytest-split setting. the (default) `duration_based_chunks`
splitting algorithm in pytest-split resulted in a 0-size chunk getting
allocated to the final group, which caused tox to error out with
InvocationError code 5.