feat: implement flags semantics for enum (#2915)
implement operators so that enums can be used like Python's `enum.Flag`
("Base class for creating enumerated constants that can be combined
using the bitwise operations without losing their Flag membership").

this also enables bitwise and/or operators for uint256, and deprecates
the builtins `bitwise_and` and `bitwise_or`

this commit also adds clampers for enums - an enum must be in bound for
the number of members the enum has (and also, any enum runtime value
must be nonzero).

Co-authored-by: tserg <8017125+tserg@users.noreply.github.com>