feat: add optimizer guardrails, refactor (#2914)
this commit factors out and generalizes the optimizer logic for
comparison operators. it clarifies the logic for dealing with boundary
cases, and cleans up the control flow in the optimizer in general. it
also adds an assembly peephole optimization to help optimize the
assembly generated by the comparator pass.

this commit also introduces a new IR keyword, "unique_symbol". it
functions as a guardrail to help ensure optimizer passes are sane. it is
a statement that codegen can insert into the IR, and sanity checks will
be performed to make sure that the statement is not optimized out, and
in some cases, to ensure that the statement shows up in the optimized
IR. it can be thought of as a very primitive effects tracking framework.

lastly, some tests are added to increase optimizer coverage.