feat: bitwise not and xor for enums (#2925)
add bitwise not and bitwise xor ops for enums. note that the
implementation uses xors with a special mask to avoid flipping upper
bits that do not belong to the enum (ex. using the `not` opcode on an
enum with 5 members does not make sense, because the 6th and higher bits
would get flipped to 1, which is out of bounds for the enum).

also, allow 0 as a valid enum value (this makes sense because storage
slots are 0 by default, and we have to assume that any storage slot with
type `MyEnum` is a valid enum value. in other words, `empty(Type)` is a
member of `Type` for any `Type`).

lastly, fix codegen for `not in` enum - it was being incorrectly handled
the same as `in`.