feat: add `code_offset=` kwarg to `create_from_factory` (#2930)
the motivation for this is that people might deploy factories with
custom prefixes (e.g. 0xFE INVALID to prevent execution of the factory
as a standalone/runtime contract, or as part of an EOF container). this
enables that usage by adding a code_offset which allows people to offset
into the factory contract address for the EXTCODECOPY.

(implementation note: the implementation does check that
`(lt code_offset (extcodesize target))`.

also ci: widen reruns delay -- try to avoid some weird deadline issues