chore: rename --debug flag to --verbose
the `--debug` flag is a bit confusing since it sounds like it turns on a
debugger mode for the compiler, when in fact it just tells the compiler
to be more verbose in its output.

this rename will also be helpful if we add a true debugger mode to the
compiler in the (near?) future.