fix: edge cases in safepow (#2983)
fix minor edge cases in safepow checks where the power or exponent is
0, -1 or 1.
 
fix edge case in safepow where raising a variable base to a power which
results in min_value() for the type results in a runtime revert (as it
is incorrectly considered OOB). for instance, before this commit,
`convert(-8, int16)**5` would revert, instead of returning
`min_value(int16)`.