docs: fix raw_call example (#2991)
fix: example for raw_call contained an error, the value for data was 0xa9059cbb. Trying the example with 0xa9059cbb could be very misleading because by using this with raw_call you ran into gas_errors (the real exception is: "vyper.exceptions.InvalidType: Expected Bytes[4] but literal can only be cast as bytes4", but this does not show up in the raw_call). So the content should have been b'\xa9\x05\x9c\xbb'. But i think it's better to show the use of method_id in this context.