fix: typechecker hotspot (#3318)
`get_possible_types_from_node` had blowup for deeply nested expressions.
this commit caches intermediate results using dynamic programming to
prune the calculation time.

note that caching has a funny interaction with the type inference
algorithm for loop iterators (i.e., try all different types for the loop
iterator, and pick the first one which does not throw an exception) -
since the loop iterator inference algorithm depends on optimistically
trying different types, some cached values can be invalid in the case
that they are generated inside of the loop iterator inference algorithm,
because the exception is caught and handled, but the state changes
(i.e., changes to the cache) are not rolled back. to solve this, this
commit introduces a simple commit/rollback mechanism to deal with state
changes in the type checker. it remains to be seen whether this
mechanism will be more generally useful in the future for other
potentially stateful type-checking operations, or if a better long-term
solution is to simply redesign how the type checker infers loop
iterators.