fix: call graph stability
the use of `set` (which does not guarantee order of its elements)
instead of `dict` (which guarantees insertion order) led to instability
across runs of the compiler between different versions of python. this
commit patches the problem by using a derivation of `dict` to track the
call graph instead of `set`.