fix: code generation for `<MyInterface>.address`
code like `<MyInterface>.address` would result being tagged with the
wrong type during codegen - `InterfaceT` instead of `AddressT`.

this was a regression introduced in 3abe588e0, and for some reason there
were no existing tests for `<MyInterface>.address` codegen. this fixes
the type tagging during codegen and adds a test to prevent this
regression in the future.