fix: typechecker nested commit/rollback
this commit fixes a bug which was introduced in 66930fdfc. when the
typechecker enters a nested loop, it can typecheck the inner loop
(committing it), and result in invalid state if the outer loop fails to
typecheck. this commit implements a checkpointing system in the
typechecker state committer so that it can handle nested changes. it
also cleans up the data structure used so that there is a single entry
point and users of the data structure do not need to think about
implementation details.

one drawback of this approach is that it can *only* handle changes to
the metadata dict. non-idempotent changes to the AST during typechecking
should then be restricted to changes to the metadata dict so that they
can register automatically with the node metadata journal.