chore: fix fuzzer test case
for cases where `a << b` is an int256 between `max_value(int256)` and
`max_value(uint256)`, the typechecker will fail with TypeMismatch
instead of OverflowException. this fixes the test to catch the case.
note that the code was correctly failing to compile, just the test was
not catching both exception types.