fix: guard against instantiation of events and hashmaps (#3341)
fix: guard against instantiation of events and hashmaps

this commit fixes a regression introduced in 046ea166d93. as a result of
removing the location kwarg from `type_from_annotation()`, events could
be instantiated in the type system (althought attempts to use them would
fail during codegen). this commit reintroduces the location kwarg, and
also streamlines some logic so that HashMaps and events can all be
handled in the same code paths.

Co-authored-by: Charles Cooper <cooper.charles.m@gmail.com>