commit message
this commit fixes two issues in constructor codegen:
- nested calls to internal functions (i.e. the constructor calls a()
  which then calls b()) were not compiling properly, only a single level
  of internal calls were tracked correctly, and
- internal functions which used immutables were being produced with the
  incorrect `dload` instead of `iload` instruction.

it fixes the first issue by simply including the code for all functions
in the generated IR for the constructor and depending on the dead code
eliminator to eliminate unused functions - instead of trying to traverse
the call graph and only including called functions.

the second issue is fixed by producing a separate compilation context
for codegen which indicates whether the function is being called from
the constrcutor or not - and depending on the context, the correct
instruction can be emitted.