fix: type annotation of Subscript nodes (#3406)
the index type was being incorrectly annotated in the frontend in the
case where the index type was not uint256. this would lead to a panic
during codegen. an example which would lead to this bug is:

```vyper
@external
def foo(xs: uint256[5], ix: uint8):
    return xs[ix + 1]  # <-- panics with uint8 != uint256
```

this commit also improves the `VyperInternalException` class so it can
include a node for prettier traceback