fix: calldatasize < 4 reverting instead of going to fallback
in the case that a selector matches calldata with less than 4 bytes, it
currently will revert instead of going to the fallback. this commit
fixes the behavior in the special case where no data besides the method
id is expected in the calldata.

note this change in behavior was originally introduced in 02339dfda.