fix: assignment when rhs is complex type and references lhs (#3410)
prior to this commit, the following code would produce a
load-after-store bug:

```vyper
@external
def foo(xs: uint256[2]) -> uint256[2]:
    ys: uint256[2] = xs
    ys = [ys[1], ys[0]]  # <-- ys[1] gets written to ys[0] before ys[0] is read
    return ys  # returns [xs[1], xs[1]]
```

this commit checks that rhs and lhs do not overlap in `parse_Assign`. if
they do, copy rhs to a temporary buffer, then copy to the lhs. this
might not be the most efficient approach, but it is safe (and it's
relatively rare to see this case in user code).

note that this commit doesn't touch `make_setter` directly because it
would need access to the memory allocator (to allocate the temporary
buffer). this could result in the fix not working in other cases where
the arguments to `make_setter` overlap. i don't think there are any
cases of those, but if there were, they would need to be fixed on a case
by case basis.