fix: calldatasize < 4 reverting instead of going to fallback (#3408)
in the case that a selector matches calldata with less than 4 bytes, it
currently will revert instead of going to the fallback. this can happen
if the selector has trailing zeroes. this commit fixes the behavior for
selectors with trailing zeroes, and improves code size and gas for those
without trailing zeroes.