fix: uninitialized immutable values (#3409)
immutable variables can be read before assignment in constructor code,
and their memory location is accessed, but that memory might not yet
be initialized. prior to this commit, its value is not necessarily
`empty(type)` since memory could have been written to ephemerally. in
particular, `create_copy_of` (and its sister, `create_from_blueprint`)
use `msize` to determine a starting location for where to copy the
target bytecode into memory, while the immutables section start is
determined using the static memory allocator. in case that `msize` is
still less than the immutables section end, `create_copy_of` can write
to the immutables section, thereby resulting in reads from the
immutables section to return garbage.

this commit fixes the issue by issuing an `iload
<immutables_section_end> - 32` before executing any initcode, which
forces `msize` to be initialized past the end of the immutables section
(and therefore, accessing an immutable before it is initialized in the
constructor will produce the "expected" `empty()` value for the
immutable).

note that a corresponding `mload` is not required for runtime code,
because vyper requires all memory variables to be instantiated at the
declaration site, so there is no way that msize can produce a pointer to
an uninitialized memory variable.